Got preview working in edit (mostly).
A few of the original DOM elements are missing (the ones that are
relates to tests). The code is still relying on global variables and
elements with unique identifier attributes (if page has a few instances
of these music exercies, then it will all fail miserable, but it is not
an issue with editing). Also introduced a spacial function and logic to
close any AudioContext elements. If these are not closed, Chrome throws
an error of going over the limit.