References Oppevara/oppevara#61, added container logic.
Made sure that exercise code would be using the correct container. This
should allow multiple exercises on the same page. It would still be
safer to force every one of those to be displayed within the IFRAME
because AudioContent limit might be enforced for the same page (needs
testing). Also madde sure that generated preview is refreshed
accordingly, if a new exercise type is chosen.