References Oppevara/oppevara#61, defined custom select widget.
Used default Select widget as a base and added custom logic to deal with
options. Gave it a unique name beccause it uses value of optgroup and
value of option to determine if it belongs there or not. Any ungrouped
options would be adde to the end. Defaults to standard widget logic if
no groups are defined.